began search implementation WIP
