+++ app.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const path = require('path')
-const session = require('express-session')
+// const session = require('express-session')
 const favicon = require('serve-favicon')
 const logger = require('morgan')
 const cookieParser = require('cookie-parser')
@@ -10,10 +10,11 @@ const passport = require( 'passport')
 const routes = require('./routes/index')
 const users = require('./routes/users')
 const books = require('./routes/books')
-const signup = require('./routes/signup')
-const signin = require('./routes/signin')
+// const signup = require('./routes/signup')
+// const signin = require('./routes/signin')
 const test = require('./routes/test')
 const models = require( './database/models' )
+const search = require('./routes/search')
 
 const app = express()
 
@@ -29,7 +30,7 @@ app.use(bodyParser.json())
 app.use(bodyParser.urlencoded({ extended: false }))
 app.use(cookieParser())
 app.use(express.static(path.join(__dirname, 'public')))
-app.use(session({ secret: 'keyboard cat', cookie: {} }))
+// app.use(session({ secret: 'keyboard cat', cookie: {} }))
 app.use(passport.initialize())
 
 app.use(passport.session())
@@ -37,9 +38,10 @@ app.use(passport.session())
 app.use('/', routes)
 app.use('/users', users)
 app.use('/books', books)
-app.use('/signup', signup)
-app.use('/signin', signin)
+// app.use('/signup', signup)
+// app.use('/signin', signin)
 app.use('/test', test)
+app.use('/search', search)
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
