+++ routes/auth.js
@@ -1,12 +1,12 @@
-var express = require('express')
-var router = express.Router()
-
-router.get('users/signin', (request, response) => {
-  response.render('signin')
-})
-
-router.get('./users/signup', (request, response) => {
-  response.render('signup')
-})
-
-module.exports = router
+// var express = require('express')
+// var router = express.Router()
+//
+// router.get('users/signin', (request, response) => {
+//   response.render('signin')
+// })
+// 
+// router.get('users/signup', (request, response) => {
+//   response.render('signup')
+// })
+//
+// module.exports = router
