+++ routes/users.js
@@ -23,4 +23,13 @@ router.post('/:id', function(request, response) {
   response.redirect('/${:id}')
 })
 
+router.get('users/signin', (request, response) => {
+  response.render('signin')
+})
+
+router.get('users/signup', (request, response) => {
+  response.render('signup')
+})
+
+
 module.exports = router
