+++ views/layout.pug
@@ -19,7 +19,7 @@ html
              a.navbar-brand(href='/') Home
            li
              a(href='/books/add') Add Book
-         form.navbar-form.navbar-left(action="/", method="get")
+         form.navbar-form.navbar-left(action="/search", method="get")
            .form-group
              input.form-control(type="text", name="search_query", placeholder="Search...")
              button.btn.btn-default(type="submit", value="Submit") Submit
